Make sure we don't emit token tree delimiters as punct
Because this breaks some proc macros.

There is actually still one place we do that: at the end of the conversion, if there are unmatched delimiters.

This has a bug: because it essentially removes unmatched delimiters, it means undoing fixup should re-add them, but I didn't do that and it's complicated to do with the way fixup is handed.